% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gazetteer.R
\name{gazetteer_open}
\alias{gazetteer_open}
\alias{gazetteer_close}
\title{Open a gazetteer snapshot}
\usage{
gazetteer_open(file)

gazetteer_close(con)
}
\arguments{
\item{file}{Path to a file written by [gazetteer_snapshot_build()].}

\item{con}{Connection from [gazetteer_open()].}
}
\value{
An open DBI connection. Close it with [gazetteer_close()].
}
\description{
Open a gazetteer snapshot
}
