% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georefapp-package.R
\docType{package}
\name{georefapp-package}
\alias{georefapp}
\alias{georefapp-package}
\title{georefapp: batch georeferencing of textual localities}
\description{
Import a table of plain-text localities, group them into distinct strings,
draw a footprint for each on a map, and export Darwin Core georeferences
backed by an append-only decision log.
}
\author{
\strong{Maintainer}: Gilles Dauby \email{gilles.dauby@ird.fr} (\href{https://orcid.org/0000-0002-9498-413X}{ORCID})

Authors:
\itemize{
  \item Gilles Dauby \email{gilles.dauby@ird.fr} (\href{https://orcid.org/0000-0002-9498-413X}{ORCID})
}

}
\keyword{internal}
