% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{store_close}
\alias{store_close}
\title{Close a project store}
\usage{
store_close(con)
}
\arguments{
\item{con}{Connection from [store_open()].}
}
\value{
`TRUE`, invisibly.
}
\description{
Close a project store
}
