% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georef-app.R
\name{georef_server}
\alias{georef_server}
\title{Server for the georeferencing application}
\usage{
georef_server()
}
\value{
A Shiny server function.
}
\description{
Server for the georeferencing application
}
